requirements.txt
pandas
numpy 
python-dotenv
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart
# -e .
